feat(revoke): backwards compatibility
Add backwards compatibility with go 1.18 or lower.
