+++ revoke/pkcs7.go
@@ -126,8 +126,8 @@ type EncryptedContentInfo struct {
 // ParsePKCS7 attempts to parse the DER encoded bytes of a
 // PKCS7 structure.
 func ParsePKCS7(raw []byte) (msg *PKCS7, err error) {
-
 	var pkcs7 initPKCS7
+
 	_, err = asn1.Unmarshal(raw, &pkcs7)
 	if err != nil {
 		return nil, WrapError(CertificateError, ParseFailed, err)
